+++ src/Sentry.Unity/MainThreadData.cs
@@ -13,15 +13,15 @@ namespace Sentry.Unity
 
         public bool? SupportsVibration { get; set; }
 
-        public Lazy<string>? DeviceType { get; set; }
+        public string? DeviceType { get; set; }
 
         public string? CpuDescription { get; set; }
 
         public string? DeviceName { get; set; }
 
-        public Lazy<string>? DeviceUniqueIdentifier { get; set; }
+        public string? DeviceUniqueIdentifier { get; set; }
 
-        public Lazy<string>? DeviceModel { get; set; }
+        public string? DeviceModel { get; set; }
 
         public int? SystemMemorySize { get; set; }
 
@@ -29,13 +29,13 @@ namespace Sentry.Unity
 
         public string? GraphicsDeviceName { get; set; }
 
-        public Lazy<string>? GraphicsDeviceVendorId { get; set; }
+        public string? GraphicsDeviceVendorId { get; set; }
 
         public string? GraphicsDeviceVendor { get; set; }
 
         public int? GraphicsMemorySize { get; set; }
 
-        public Lazy<bool>? GraphicsMultiThreaded { get; set; }
+        public bool? GraphicsMultiThreaded { get; set; }
 
         public string? NpotSupport { get; set; }
 
@@ -55,15 +55,17 @@ namespace Sentry.Unity
 
         public int? GraphicsShaderLevel { get; set; }
 
-        public Lazy<bool>? IsDebugBuild { get; set; }
+        public bool? IsDebugBuild { get; set; }
 
         public string? InstallMode { get; set; }
 
-        public Lazy<string>? TargetFrameRate { get; set; }
+        public string? TargetFrameRate { get; set; }
 
-        public Lazy<string>? CopyTextureSupport { get; set; }
+        public string? CopyTextureSupport { get; set; }
 
-        public Lazy<string>? RenderingThreadingMode { get; set; }
+        public string? RenderingThreadingMode { get; set; }
+
+        public DateTimeOffset? StartTime { get; set; }
 
         public bool IsMainThread()
             => MainThreadId.HasValue && Thread.CurrentThread.ManagedThreadId == MainThreadId;